+++ src/api/util.rs
@@ -2,19 +2,19 @@ use super::*;
 
 /// Convert a 20 byte Account ID to an r-address
 #[inline(always)]
-pub fn util_raddr(raddr_out: &mut [u8], accid: &AccountId) -> Result<i64> {
+pub fn util_raddr(raddr_out: &mut [u8], accid: &[u8]) -> Result<i64> {
     buf_write_read(raddr_out, accid, _c::util_raddr)
 }
 
 /// Convert an r-address into a 20 byte Account ID
 #[inline(always)]
-pub fn util_accid(accid_out: &mut AccountId, raddr_in: &[u8]) -> Result<i64> {
+pub fn util_accid(accid_out: &mut [u8], raddr_in: &[u8]) -> Result<i64> {
     buf_write_read(accid_out, raddr_in, _c::util_accid)
 }
 
 /// Verify a cryptographic signature
 #[inline(always)]
-pub fn util_verify(payload: &[u8], signature: &[u8], publickey: &PublicKey) -> bool {
+pub fn util_verify(payload: &[u8], signature: &[u8], publickey: &[u8]) -> bool {
     let res = buf_3_read(payload, signature, publickey, _c::util_verify);
 
     match res {
@@ -26,13 +26,13 @@ pub fn util_verify(payload: &[u8], signature: &[u8], publickey: &PublicKey) -> b
 
 /// Compute an sha512-half over some data
 #[inline(always)]
-pub fn util_sha512h(hash_out: &mut Hash, data_in: &[u8]) -> Result<i64> {
+pub fn util_sha512h(hash_out: &mut [u8], data_in: &[u8]) -> Result<i64> {
     buf_write_read(hash_out, data_in, _c::util_sha512h)
 }
 
 /// Compute a serialized keylet of a given type
 #[inline(always)]
-pub fn util_keylet(keylet: &mut Keylet, keylet_type: KeyletType) -> Result<i64> {
+pub fn util_keylet(keylet: &mut [u8], keylet_type: KeyletType) -> Result<i64> {
     let write_ptr = keylet.as_mut_ptr() as _;
     let write_len = keylet.len() as _;