+++ src/lib.rs
@@ -22,7 +22,7 @@
     unreachable_pub
 )]
 #![doc(test(attr(deny(warnings))))]
-#![doc(html_root_url = "https://docs.rs/xrpl-hooks/0.5.0")]
+#![doc(html_root_url = "https://docs.rs/xrpl-hooks/0.6.0")]
 
 mod macros;