+++ HISTORY.md
@@ -4,7 +4,8 @@
 
 * Added `require` and `is_buffer_equal_20`, `is_buffer_equal_32` helpers
 * Added `drops_to_amount` helper
-  
+* Revert strict types
+
 ## 0.5.0 (2024-9-22)
 
 * Refactor API functions to use specific types